Evaluate word2vec model with similarity metric
For every run-running-spin-spinning relationship in our set of relationships,
we measure the similarity between (x_run - x_running) and (x_spin - x_spinning)
where x is the word vector of the subscript. We take the mean of all these
similarities as a one number metric for how our relationship is performing.