Train Biomed word2vec model using generators
Generators yield sentences one-by-one as opposed to lists which store all the
sentences in memory. As such generators are more suitable when training with
large amounts of data. This commit is in the step towards using more generators
but at the final step, sentences are still stored all in memory. More work is
required.